Optimize fill loops

+++ examples/export_image/mod.rs
@@ -8,7 +8,7 @@ pub fn save_raster_as_image(filename: &str, raster: GreyscaleRaster){
 
     for x in 0..width {
         for y in 0..height {
-            let pix = raster.get_pixel(x as u32, y as u32);
+            let pix = raster.get_pixel(x, y);
             let i_pix = 255 - pix;
             
             imgbuf.put_pixel(x as u32, y as u32, Rgba([i_pix, i_pix, i_pix, 255]));
